Snapshot from 6/30/2023
Changed some datatypes from double to single.
Began transitioning to Datastore-based image and intensity profile data storage.
Fixed several issues with the queue on reanalysis.

+++ src/+sbsense/+improc/analyzeComposite.m
@@ -32,7 +32,7 @@ function [Yc, Yr, peakData, estimatedLaserIntensity, ...
     intprofs = NaN(numIPpoints, numChannels);
     fitprofs = NaN(numIPpoints, numChannels);
     cfitBoundses = NaN(2, numChannels);
-    peakSearchBounds = double(peakSearchBounds) + [1 -1]; % TODO: Eliminate need for offset
+    peakSearchBounds = uint16(int32(peakSearchBounds) + [1 -1]); % TODO: Eliminate need for offset
     if ~successTF
         fprintf(f, '[analyzeComposite] Estimation of laser intensity was unsuccessful. Returning without performing further analysis.\n');
         return;
